+++ src/schema/statetrace/statetrace_test.go
@@ -1,4 +1,4 @@
-package ledger
+package statetrace
 
 import (
 	"context"
